perf: short lodash import size, rename progress bar file
