+++ package.json
@@ -35,6 +35,7 @@
     "jsdom": "^20.0.0",
     "postcss": "^8.4.16",
     "prettier": "^2.7.1",
+    "rollup-plugin-visualizer": "^5.8.0",
     "start-server-and-test": "^1.14.0",
     "tailwindcss": "^3.1.8",
     "vite": "^3.0.4",
