+++ src/includes/progress-bar.js
