+++ vite.config.js
@@ -3,6 +3,7 @@ import { fileURLToPath, URL } from "node:url";
 import { defineConfig } from "vite";
 import { VitePWA } from "vite-plugin-pwa";
 import vue from "@vitejs/plugin-vue";
+// import { visualizer } from "rollup-plugin-visualizer";
 
 // https://vitejs.dev/config/
 export default defineConfig({
@@ -28,6 +29,7 @@ export default defineConfig({
         globPatterns: ["**/*.{js,css,html,png,jpg}"],
       },
     }),
+    // visualizer({ open: true }),
   ],
   resolve: {
     alias: {
